+++ src/FormsySelect.jsx
@@ -10,6 +10,9 @@ const FormsySelect = React.createClass({
     name: React.PropTypes.string.isRequired,
     onChange: React.PropTypes.func,
     requiredError: React.PropTypes.string,
+    validationError: React.PropTypes.string,
+    validationErrors: React.PropTypes.object,
+    validations: React.PropTypes.oneOfType([React.PropTypes.string, React.PropTypes.object]),
     value: React.PropTypes.any,
   },
 
@@ -34,7 +37,13 @@ const FormsySelect = React.createClass({
   setMuiComponentAndMaybeFocus: setMuiComponentAndMaybeFocus,
 
   render() {
-    let { value, ...rest } = this.props;
+    let { value } = this.props;
+
+    const { validations, // eslint-disable-line no-unused-vars
+        validationError, // eslint-disable-line no-unused-vars
+        validationErrors, // eslint-disable-line no-unused-vars
+      ...rest } = this.props;
+
     value = this.state.hasChanged ? this.getValue() : value;
     const { requiredError } = this.props;
     const { isRequired, isPristine, isValid, isFormSubmitted } = this;