+++ src/FormsyTime.jsx
@@ -9,6 +9,9 @@ const FormsyTime = React.createClass({
     defaultTime: React.PropTypes.object,
     name: React.PropTypes.string.isRequired,
     onChange: React.PropTypes.func,
+    validationError: React.PropTypes.string,
+    validationErrors: React.PropTypes.object,
+    validations: React.PropTypes.oneOfType([React.PropTypes.string, React.PropTypes.object]),
     value: React.PropTypes.object,
   },
 
@@ -33,6 +36,9 @@ const FormsyTime = React.createClass({
   render() {
     const {
       defaultTime, // eslint-disable-line no-unused-vars
+      validations, // eslint-disable-line no-unused-vars
+      validationError, // eslint-disable-line no-unused-vars
+      validationErrors, // eslint-disable-line no-unused-vars
       ...rest,
     } = this.props;